pyproject.toml
[project]
name = "cross-validation-for-time-series-data"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.3",
    "pmdarima>=2.1.1",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
